+++ shared/src/commonMain/kotlin/site/remlit/snowdrop/component/Emoji.kt
@@ -5,6 +5,12 @@ import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.size
 import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.PlainTooltip
+import androidx.compose.material3.Text
+import androidx.compose.material3.TooltipAnchorPosition
+import androidx.compose.material3.TooltipBox
+import androidx.compose.material3.TooltipDefaults
+import androidx.compose.material3.rememberTooltipState
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.layout.ContentScale
@@ -29,7 +35,9 @@ const val bigEmojiSize = 40
 fun Emoji(
 	emoji: Emoji,
 	fill: Boolean = false,
-	big: Boolean = false
+	big: Boolean = false,
+
+	showTooltip: Boolean = true
 ) {
 	val size = if (big) bigEmojiSize.dp else emojiSize.dp
 
@@ -46,12 +54,30 @@ fun Emoji(
 		)
 	}
 
-	// todo: reduced motion should make this use emoji.staticUrl
-	KamelImage(
-		resource = { asyncPainterResource(emoji.url) },
-		contentDescription = emoji.shortcode,
-		contentScale = ContentScale.Fit,
-		onLoading = { fallback() },
-		modifier = Modifier.emojiSize(),
-	)
+	@Composable
+	fun renderContent() {
+		// todo: reduced motion should make this use emoji.staticUrl
+		KamelImage(
+			resource = { asyncPainterResource(emoji.url) },
+			contentDescription = emoji.shortcode,
+			contentScale = ContentScale.Fit,
+			onLoading = { fallback() },
+			onFailure = { fallback() },
+			modifier = Modifier.emojiSize(),
+		)
+	}
+
+	if (showTooltip) {
+		val tooltipState = rememberTooltipState()
+		val positionProvider = TooltipDefaults.rememberTooltipPositionProvider(
+			TooltipAnchorPosition.Above
+		)
+
+		TooltipBox(
+			modifier = Modifier,
+			positionProvider = positionProvider,
+			state = tooltipState,
+			tooltip = { PlainTooltip { Text(":${emoji.shortcode}:") } },
+		) { renderContent() }
+	} else renderContent()
 }